renamed Convention Class Added Controller name customization feature
